adicao de autenticacao de usuario no controller e views

+++ resources/views/produto/index.blade.php
@@ -39,10 +39,12 @@
                          {{ $produto->titulo }}
                       </a>
                   @endif
-                  {{Form::open(['route'=>['produtos.destroy',$produto->id],'method'=>'DELETE'])}}
+                  @if(Auth::check())
+                      {{Form::open(['route'=>['produtos.destroy',$produto->id],'method'=>'DELETE'])}}
                         <a href="{{url('produtos/'.$produto->id.'/edit')}}" class="btn btn-default">Editar</a>
-                    {{Form::submit('Excluir',['class'=> 'btn btn-default'])}}
-                  {{Form::close()}}
+                        {{Form::submit('Excluir',['class'=> 'btn btn-default'])}}
+                      {{Form::close()}}
+                  @endif
               </div>
         @endforeach
     </div>
